updated namespaces for solutions and projects
