+++ samples/api-only/Microsoft.Restier.Samples.Apis.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Restier.Samples.Chinook.Api", "chinook\Microsoft.Restier.Samples.Chinook.Api\Microsoft.Restier.Samples.Chinook.Api.csproj", "{A0920D03-AB1D-4F56-ABF7-2C694E100F09}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Restier.Samples.Chinook.Tests.Api", "chinook\Microsoft.Restier.Samples.Chinook.Tests.Api\Microsoft.Restier.Samples.Chinook.Tests.Api.csproj", "{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A0920D03-AB1D-4F56-ABF7-2C694E100F09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0920D03-AB1D-4F56-ABF7-2C694E100F09}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0920D03-AB1D-4F56-ABF7-2C694E100F09}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0920D03-AB1D-4F56-ABF7-2C694E100F09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
