--- samples/api-only/chinook/ChinookRESTierAPI/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="ChinookRESTierAPI.WebApiApplication" Language="C#" %>
