--- samples/api-only/chinook/ChinookRESTierAPI/Models/Employee.cs
@@ -1,69 +1,0 @@
-namespace ChinookRESTierAPI.Models
-{
-    using System;
-    using System.Collections.Generic;
-    using System.ComponentModel.DataAnnotations;
-    using System.ComponentModel.DataAnnotations.Schema;
-
-    [Table("Employee")]
-    public sealed class Employee
-    {
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
-        public Employee()
-        {
-            Customers = new HashSet<Customer>();
-            Employee1 = new HashSet<Employee>();
-        }
-
-        public int EmployeeId { get; set; }
-
-        [Required]
-        [StringLength(20)]
-        public string LastName { get; set; }
-
-        [Required]
-        [StringLength(20)]
-        public string FirstName { get; set; }
-
-        [StringLength(30)]
-        public string Title { get; set; }
-
-        public int? ReportsTo { get; set; }
-
-        public DateTime? BirthDate { get; set; }
-
-        public DateTime? HireDate { get; set; }
-
-        [StringLength(70)]
-        public string Address { get; set; }
-
-        [StringLength(40)]
-        public string City { get; set; }
-
-        [StringLength(40)]
-        public string State { get; set; }
-
-        [StringLength(40)]
-        public string Country { get; set; }
-
-        [StringLength(10)]
-        public string PostalCode { get; set; }
-
-        [StringLength(24)]
-        public string Phone { get; set; }
-
-        [StringLength(24)]
-        public string Fax { get; set; }
-
-        [StringLength(60)]
-        public string Email { get; set; }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public ICollection<Customer> Customers { get; set; }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public ICollection<Employee> Employee1 { get; set; }
-
-        public Employee Employee2 { get; set; }
-    }
-}
