+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/App_Start/WebApiConfig.cs
@@ -1,10 +1,10 @@
 ﻿using System.Web.Http;
-using ChinookRESTierAPI.Models;
 using Microsoft.AspNet.OData.Extensions;
 using Microsoft.Restier.AspNet.Batch;
 using Microsoft.Restier.EntityFramework;
+using Microsoft.Restier.Samples.Chinook.Api.Models;
 
-namespace ChinookRESTierAPI
+namespace Microsoft.Restier.Samples.Chinook.Api
 {
     public static class WebApiConfig
     {
@@ -14,7 +14,7 @@ namespace ChinookRESTierAPI
             await config.MapRestierRoute<EntityFrameworkApi<ChinookContext>>(
                 "Chinook",
                 "api/",
-                new RestierBatchHandler(GlobalConfiguration.DefaultServer));
+                new RestierBatchHandler(GlobalConfiguration.DefaultServer)).ConfigureAwait(false);
         }
     }
 }
