+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Microsoft.Restier.Samples.Chinook.Api.WebApiApplication" Language="C#" %>
