+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Microsoft.Restier.Samples.Chinook.Api.csproj
@@ -11,8 +11,8 @@
     <ProjectTypeGuids>{349c5851-65df-11da-9384-00065b846f21};{fae04ec0-301f-11d3-bf4b-00c04f79efbc}</ProjectTypeGuids>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ChinookRESTierAPI</RootNamespace>
-    <AssemblyName>ChinookRESTierAPI</AssemblyName>
+    <RootNamespace>Microsoft.Restier.Samples.Chinook.Api</RootNamespace>
+    <AssemblyName>Microsoft.Restier.Samples.Chinook.Api</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <UseIISExpress>true</UseIISExpress>
     <Use64BitIISExpress />
