+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Models/Album.cs
@@ -1,24 +1,29 @@
-namespace ChinookRESTierAPI.Models
+namespace Microsoft.Restier.Samples.Chinook.Api.Models
 {
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.ComponentModel.DataAnnotations.Schema;
 
-    [Table("Artist")]
-    public sealed class Artist
+    [Table("Album")]
+    public sealed class Album
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
-        public Artist()
+        public Album()
         {
-            Albums = new HashSet<Album>();
+            Tracks = new HashSet<Track>();
         }
+
+        public int AlbumId { get; set; }
+
+        [Required]
+        [StringLength(160)]
+        public string Title { get; set; }
 
         public int ArtistId { get; set; }
 
-        [StringLength(120)]
-        public string Name { get; set; }
+        public Artist Artist { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public ICollection<Album> Albums { get; set; }
+        public ICollection<Track> Tracks { get; set; }
     }
 }
