+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Models/Customer.cs
@@ -1,0 +1,61 @@
+namespace Microsoft.Restier.Samples.Chinook.Api.Models
+{
+    using System.Collections.Generic;
+    using System.ComponentModel.DataAnnotations;
+    using System.ComponentModel.DataAnnotations.Schema;
+
+    [Table("Customer")]
+    public sealed class Customer
+    {
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
+        public Customer()
+        {
+            Invoices = new HashSet<Invoice>();
+        }
+
+        public int CustomerId { get; set; }
+
+        [Required]
+        [StringLength(40)]
+        public string FirstName { get; set; }
+
+        [Required]
+        [StringLength(20)]
+        public string LastName { get; set; }
+
+        [StringLength(80)]
+        public string Company { get; set; }
+
+        [StringLength(70)]
+        public string Address { get; set; }
+
+        [StringLength(40)]
+        public string City { get; set; }
+
+        [StringLength(40)]
+        public string State { get; set; }
+
+        [StringLength(40)]
+        public string Country { get; set; }
+
+        [StringLength(10)]
+        public string PostalCode { get; set; }
+
+        [StringLength(24)]
+        public string Phone { get; set; }
+
+        [StringLength(24)]
+        public string Fax { get; set; }
+
+        [Required]
+        [StringLength(60)]
+        public string Email { get; set; }
+
+        public int? SupportRepId { get; set; }
+
+        public Employee Employee { get; set; }
+
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
+        public ICollection<Invoice> Invoices { get; set; }
+    }
+}
