+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Models/Genre.cs
@@ -1,19 +1,19 @@
-namespace ChinookRESTierAPI.Models
+namespace Microsoft.Restier.Samples.Chinook.Api.Models
 {
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.ComponentModel.DataAnnotations.Schema;
 
-    [Table("MediaType")]
-    public sealed class MediaType
+    [Table("Genre")]
+    public sealed class Genre
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
-        public MediaType()
+        public Genre()
         {
             Tracks = new HashSet<Track>();
         }
 
-        public int MediaTypeId { get; set; }
+        public int GenreId { get; set; }
 
         [StringLength(120)]
         public string Name { get; set; }
