+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("ChinookRESTierAPI")]
+[assembly: AssemblyTitle("Microsoft.Restier.Samples.Chinook.Api")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("ChinookRESTierAPI")]
+[assembly: AssemblyProduct("Microsoft.Restier.Samples.Chinook.Api")]
 [assembly: AssemblyCopyright("Copyright ©  2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
