+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/Web.config
