+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Api/packages.config
