+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Tests.Api/Microsoft.Restier.Samples.Chinook.Tests.Api.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{41BAC0CC-C2B0-47B8-91C6-8E0E9FCBCF44}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ChinookRESTierAPI.Tests</RootNamespace>
-    <AssemblyName>ChinookRESTierAPI.Tests</AssemblyName>
+    <RootNamespace>Microsoft.Restier.Samples.Chinook.Tests.Api</RootNamespace>
+    <AssemblyName>Microsoft.Restier.Samples.Chinook.Tests.Api</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
