+++ samples/api-only/chinook/Microsoft.Restier.Samples.Chinook.Tests.Api/packages.config
